add feature test element and rename variables applicably
