+++ src/__tests__/userCanCreateTimeRecording.spec.js
@@ -1,16 +1,16 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import axios from 'axios';
-import { DisplayTimeRecording } from "../Components/DisplayTimeRecording";
+import SaveTimeRecording from "../components/SaveTimeRecording";
 
 jest.mock('axios')
 
-describe('< DisplayTimeRecording/>', () => {
+describe('< SaveTimeRecording/>', () => {
   it('should post times to backend when clicked', () => {
     const axiosSpy = jest.spyOn(axios, 'post');
     shallow(
-      <DisplayTimeRecording/>
+      <SaveTimeRecording/>
     )
     expect(axiosSpy).toBeCalled();
-  }) 
+  })
 })
